# specs/features/specs-management/api.yaml
openapi: 3.1.0
info:
  title: Specs Management API
  version: 0.1.0
  description: |
    仕様書 (specs/ 配下の md / yaml) を管理する API。
    `{id}` は specs/ からの相対パス (例: features/user-login/spec.md)。
servers:
  - url: /api
paths:
  /specs:
    get:
      summary: 仕様書一覧の取得
      operationId: listSpecs
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  specs:
                    type: array
                    items:
                      $ref: "#/components/schemas/Spec"
  /specs/{id}:
    parameters:
      - $ref: "#/components/parameters/SpecId"
    get:
      summary: 仕様書詳細の取得
      operationId: getSpec
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  spec:
                    $ref: "#/components/schemas/Spec"
                  content:
                    type: string
        "404":
          $ref: "#/components/responses/Error"
    put:
      summary: 仕様書の更新
      description: api.yaml の場合は保存前に OpenAPI スキーマ検証を行う。
      operationId: updateSpec
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
              required:
                - content
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
    delete:
      summary: 仕様書の削除
      operationId: deleteSpec
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "404":
          $ref: "#/components/responses/Error"
  /features:
    post:
      summary: feature の作成 (spec.md / api.yaml を生成)
      operationId: createFeature
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
              required:
                - name
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  created:
                    type: array
                    items:
                      type: string
        "400":
          $ref: "#/components/responses/Error"
  /features/{feature}/screens:
    parameters:
      - $ref: "#/components/parameters/Feature"
    post:
      summary: 画面の作成 (screens/S-00n.md)
      operationId: createScreen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
              required:
                - name
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  created:
                    type: string
        "400":
          $ref: "#/components/responses/Error"
  /features/{feature}/screens/order:
    parameters:
      - $ref: "#/components/parameters/Feature"
    put:
      summary: 画面の並び替え (frontmatter の order を更新)
      operationId: reorderScreens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                order:
                  type: array
                  items:
                    type: string
              required:
                - order
      responses:
        "200":
          $ref: "#/components/responses/Ok"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
components:
  parameters:
    SpecId:
      name: id
      in: path
      required: true
      description: specs/ からの相対パス
      schema:
        type: string
    Feature:
      name: feature
      in: path
      required: true
      schema:
        type: string
  responses:
    Ok:
      description: OK
      content:
        application/json:
          schema:
            type: object
            properties:
              ok:
                type: boolean
    Error:
      description: エラー
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Spec:
      type: object
      properties:
        id:
          type: string
        feature:
          type: string
        file:
          type: string
        type:
          type: string
          enum: [feature, api, screen, term, model]
        title:
          type: string
        status:
          type: string
        order:
          type: integer
      required:
        - id
        - file
        - type
    Error:
      type: object
      properties:
        error:
          type: string
      required:
        - error